+++ daemon/internal/runconfig/fixtures/windows/container_config_1_24.json
